requirements.txt
# Core dependencies
pandas==2.1.4
openpyxl==3.1.2

# Google Cloud AI
google-cloud-aiplatform==1.38.1
google-cloud-firestore==2.14.0

# Web framework (for Cloud Run deployment)
flask==3.0.0
gunicorn==21.2.0

# Utilities
python-dotenv==1.0.0

